+++ commit/src/mmcs.rs
@@ -31,7 +31,7 @@ pub trait Mmcs<T>: Clone {
     /// Opens a batch of rows from committed matrices
     /// returns `(openings, proof)`
     /// where `openings` is a vector whose `i`th element is the `j`th row of the ith matrix `M[i]`,
-    /// and `j = index >> (il(max_height) - log2_ceil(M[i].height))`.
+    /// and `j = index >> (log2_ceil(max_height) - log2_ceil(M[i].height))`.
     fn open_batch(
         &self,
         index: usize,
