+++ symmetric/src/lib.rs
@@ -5,12 +5,14 @@
 extern crate alloc;
 
 mod compression;
+mod hash;
 mod hasher;
 mod permutation;
 mod serializing_hasher;
 mod sponge;
 
 pub use compression::*;
+pub use hash::*;
 pub use hasher::*;
 pub use permutation::*;
 pub use serializing_hasher::*;
